User authorization completed
userIDs required now for note creation; middleware checks

+++ backend/src/controllers/notes-controller.ts
@@ -2,10 +2,14 @@ import {RequestHandler} from "express";
 import noteSchema from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import {assertisDefined} from "../util/assertisDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await noteSchema.find().exec();
+    assertisDefined(authenticatedUserId); // Check if authenticatedUserId is defined
+    const notes = await noteSchema.find({userId: authenticatedUserId}).exec();
     // If notes found, return them
     res.status(200).json(notes);
   } catch (error) {
@@ -15,7 +19,9 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertisDefined(authenticatedUserId); // Check if authenticatedUserId is defined
     // Check if note ID is valid
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
@@ -24,6 +30,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
     // If note not found, throw error
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+    // Check if note belongs to authenticated user
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
     }
     // If note found, return it
     res.status(200).json(note);
@@ -42,11 +52,14 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
   const text = req.body.text;
   const title = req.body.title;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertisDefined(authenticatedUserId); // Check if authenticatedUserId is defined
     if (!title) {
       throw createHttpError(400, 'Title is missing from note');
     }
     const newNote = await noteSchema.create({
+      userId: authenticatedUserId,
       text: text,
       title: title,
     });
@@ -70,7 +83,9 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
   const noteId = req.params.noteId;
   const newText = req.body.text;
   const newTitle = req.body.title;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertisDefined(authenticatedUserId); // Check if authenticatedUserId is defined
     // Check if note ID is valid
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
@@ -82,6 +97,9 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     // If note not found, throw error
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot update this note');
     }
     // If note found, update it
     note.text = newText;
@@ -97,7 +115,9 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertisDefined(authenticatedUserId); // Check if authenticatedUserId is defined
     // Check if note ID is valid
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
@@ -106,6 +126,9 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
     // If note not found, throw error
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot delete this note');
     }
     // If note found, delete it
     await note.deleteOne();
